jsonstreams: Add type stubs
This is a typing stub file for jsonstreams. I would really prefer to
include the type annotations directly in the file itself, but since
python 2.7 and 3.3 are support requirements that's not possible.

This also adds a new tox target to test the module and the boilerplate
to setup.py to install these.

Fixes #6