jsonstreams: Fix BaseWriter.comma not being set for pretty printing
Currently the Writer classes implement BaseWriter.comma as a property,
that compares self.writer to self._write_no_comma. Which works fine for
the non-pretty printing path, but doesn't work at all with pretty
printing.

This patches changes self.comma to an attribute of the writer, and has
set_comma modify it.  This does slightly increase the memory usage of
the BaseWriter class, but simplifies the implementation because it's set
in the set_comma method.

Fixes #12