jsonstreams: use slots for Writers.
This adds the __slots__ attribute to the Private Writer classes, which
has the potential to reduce their memory consumption, since there is the
possibility of opening a large number of these objects.

This does not apply them to the Object and Array classes because it
makes their attributes read only, which has the side effect of making
the close method's trick of replacing methods with functions that raise
exceptions impossible and would force use to use if's. I don't think
that's a trade-off I want to make.