Add api to control closing the fd
The Stream class has always assumed that it has ownership of the
file-like object it is writing into. This isn't (and can't be) true in
some cases, like writing to stdout. This commit adds a new argument to
the Stream class, close_fd, which signals to the Stream class whether it
owns the fd or not, and thus whether it should close it.

Fixes: #23