Flush the fd on close
This would have happened in fd.close(), and if the fd happens to be in
in use multiple places then you can get mixed output. While it probably
makes sense if you control the fd to create a locking wrapper, this is
difficult for things like `sys.stdout`, which would require monkey
patching to guard with locks if using any external modules (including
the standard library).

Fixes: #24